refactor: move logic to conflict check wrapper
